# SYCL: Add non contiguous support in RMS_NORM and NORM kernels