--- a/ggml/src/ggml-sycl/common.hpp
+++ b/ggml/src/ggml-sycl/common.hpp
@@ -475,15 +475,11 @@ static __dpct_inline__ float warp_reduce_max(float x,
 /* Helper for Computing the linear offset into an 4-dimensional ggml_tensor given
 per-dimension sizes, strides, and indices */
 template<int N>
-static __dpct_inline__ size_t calculate_offset(const std::array<int, N> & dims, const std::array<int, N> & strides, const std::array<int, N> & indices) {
+static __dpct_inline__ size_t calculate_offset(const std::array<int, N> & strides, const std::array<int, N> & indices) {
     size_t offset = 0;
 #pragma unroll
     for (int i = 0; i < N; i++) {
         auto index_i = indices[i];
-        // Handle wrap-around for indices that exceed dimensions
-        if (indices[i] >= dims[i]) {
-            index_i = indices[i] % dims[i];
-        }
         offset += strides[i] * index_i;
     }
     return offset;