Added non contiguous support in RMS_NORM and NORM kernels. `test-backend-ops` seems to pass with this change.