@qnixsynapse 2025-05-19
It now seems to pass with ne[0] = 1920
```
 NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000000): OK
  NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000000): OK
  NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000000): OK
  NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000000): OK
  NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000001): OK
  NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000001): OK
  NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000001): OK
  NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000001): OK
  NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000100): OK
  NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000100): OK
  NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000100): OK
  NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000100): OK
  NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.100000): OK
  NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.100000): OK
  NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.100000): OK
  NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.100000): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000000): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000000): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000000): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000000): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000001): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000001): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000001): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000001): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.000100): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.000100): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.000100): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.000100): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=0,eps=0.100000): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=0,eps=0.100000): OK
  RMS_NORM(type=f32,ne=[64,5,4,3],v=1,eps=0.100000): OK
  RMS_NORM(type=f32,ne=[1920,5,4,3],v=1,eps=0.100000): OK
  5535/5535 tests passed
  Backend SYCL0: OK
```