@qnixsynapse 2025-05-19
> Were you able to measure the impact on performance for this change? If it has one we may want to introduce different paths for contiguous and non-contiguous cases.
> 
> I'll try to check on my side for some relevant sizes at some point.

Doesn't seem much different from master:

[llama.cpp][master]$ build/bin/llama-bench -ngl 99 -m ~/Downloads/Weights/pythia-1.4b-q4_0.gguf 
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gptneox 1.4B Q4_0              | 786.31 MiB |     1.41 B | SYCL       |  99 |           pp512 |       4202.02 ± 7.96 |
| gptneox 1.4B Q4_0              | 786.31 MiB |     1.41 B | SYCL       |  99 |           tg128 |         46.34 ± 0.25 |

build: 92ecdcc0 (5423)

[llama.cpp][sycl/non_cont_norms]$ build/bin/llama-bench -ngl 99 -m ~/Downloads/Weights/pythia-1.4b-q4_0.gguf 
| model                          |       size |     params | backend    | ngl |            test |                  t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | --------------: | -------------------: |
| gptneox 1.4B Q4_0              | 786.31 MiB |     1.41 B | SYCL       |  99 |           pp512 |       4202.28 ± 8.47 |
| gptneox 1.4B Q4_0              | 786.31 MiB |     1.41 B | SYCL       |  99 |           tg128 |         46.25 ± 0.26 |

build: This PR