@Alcpz 2025-05-20
I ran a couple of models and  don't see any regressions there. It seems the changes don't affect the existing codepath:

| model                    | device        | test   |   avg_ts 5424 |   avg_ts 5429 |   speedup |
|:-------------------------|:--------------|:-------|--------------:|--------------:|----------:|
| qwen2 1.5B Q4_0          | Arc A770      | pp512  |     4042.61   |     4020.68   |      0.99 |
| qwen2 1.5B Q4_0          | Arc A770      | tg128  |       46.9076 |       46.919  |      1    |
| qwen2 1.5B Q4_K - Medium | Arc A770      | pp512  |     4418.48   |     4408.21   |      1    |
| qwen2 1.5B Q4_K - Medium | Arc A770      | tg128  |       46.3589 |       45.9716 |      0.99 |
| llama 7B Q4_0            | Arc A770      | pp512  |     1460.58   |     1460.19   |      1    |
| llama 7B Q4_0            | Arc A770      | tg128  |       34.986  |       34.7013 |      0.99 |
| llama 7B Q4_K - Medium   | Arc A770      | pp512  |     1718.93   |     1717.47   |      1    |
| llama 7B Q4_K - Medium   | Arc A770      | tg128  |       32.7239 |       32.8426 |      1    |
| gemma2 2B Q4_K - Medium  | Arc A770      | pp512  |     3584.07   |     3581.23   |      1    |
| gemma2 2B Q4_K - Medium  | Arc A770      | tg128  |       39.4594 |       39.3022 |      1    |
| phi3 3B Q4_0             | Arc A770      | pp512  |     2153.61   |     2156.82   |      1    |
| phi3 3B Q4_0             | Arc A770      | tg128  |       40.3495 |       40.2226 |      1    |
| phi3 3B Q4_K - Medium    | Arc A770      | pp512  |     2492.66   |     2491.55   |      1    |
| phi3 3B Q4_K - Medium    | Arc A770      | tg128  |       34.7158 |       34.788  |      1    |
| llama 34B Q6_K           | Arc A770      | pp512  |     1023.89   |     1021.92   |      1    |
| llama 34B Q6_K           | Arc A770      | tg128  |       15.2625 |       15.2357 |      1    |
| llama 8B Q4_K - Medium   | Arc A770      | pp512  |     1615.01   |     1616.76   |      1    |
| llama 8B Q4_K - Medium   | Arc A770      | tg128  |       28.8256 |       28.7515 |      1    |