@Rbiessy 2025-05-20
I looked at some common sizes used with `rms_norm` and ran a few of them with `test-backend-ops perf` on LNL iGPU:
Master:
```
  NORM(type=f32,ne=[4096,1,1,1],v=0,eps=0.100000):                     57344 runs -    18.59 us/run -       32 kB/run -    1.64 GB/s
  NORM(type=f32,ne=[4096,512,1,1],v=0,eps=0.100000):                    6147 runs -   229.37 us/run -    16384 kB/run -   68.12 GB/s
  RMS_NORM(type=f32,ne=[1536,1,1,1],v=0,eps=0.100000):                 81910 runs -    12.92 us/run -       12 kB/run -    0.89 GB/s
  RMS_NORM(type=f32,ne=[1536,512,1,1],v=0,eps=0.100000):               10924 runs -   132.90 us/run -     6144 kB/run -   44.09 GB/s
  RMS_NORM(type=f32,ne=[3072,1,1,1],v=0,eps=0.100000):                 98292 runs -    10.40 us/run -       24 kB/run -    2.20 GB/s
  RMS_NORM(type=f32,ne=[3072,512,1,1],v=0,eps=0.100000):                5462 runs -   191.00 us/run -    12288 kB/run -   61.37 GB/s
  RMS_NORM(type=f32,ne=[4096,1,1,1],v=0,eps=0.100000):                 98292 runs -    10.81 us/run -       32 kB/run -    2.82 GB/s
  RMS_NORM(type=f32,ne=[4096,512,1,1],v=0,eps=0.100000):                6147 runs -   225.46 us/run -    16384 kB/run -   69.32 GB/s
```

PR:
```
  NORM(type=f32,ne=[4096,1,1,1],v=0,eps=0.100000):                     57344 runs -    18.88 us/run -       32 kB/run -    1.62 GB/s
  NORM(type=f32,ne=[4096,512,1,1],v=0,eps=0.100000):                    6147 runs -   236.88 us/run -    16384 kB/run -   65.96 GB/s
  RMS_NORM(type=f32,ne=[1536,1,1,1],v=0,eps=0.100000):                 73719 runs -    14.56 us/run -       12 kB/run -    0.79 GB/s
  RMS_NORM(type=f32,ne=[1536,512,1,1],v=0,eps=0.100000):               10924 runs -   140.06 us/run -     6144 kB/run -   41.84 GB/s
  RMS_NORM(type=f32,ne=[3072,1,1,1],v=0,eps=0.100000):                 98292 runs -    10.37 us/run -       24 kB/run -    2.21 GB/s
  RMS_NORM(type=f32,ne=[3072,512,1,1],v=0,eps=0.100000):                5462 runs -   192.81 us/run -    12288 kB/run -   60.79 GB/s
  RMS_NORM(type=f32,ne=[4096,1,1,1],v=0,eps=0.100000):                 98292 runs -    10.91 us/run -       32 kB/run -    2.80 GB/s
  RMS_NORM(type=f32,ne=[4096,512,1,1],v=0,eps=0.100000):                6147 runs -   228.09 us/run -    16384 kB/run -   68.52 GB/s
```

There is a bit of noise but overall there is some overhead. From what we've seen `*_NORM` is low enough in the list of kernels sorted by total time spent on the device so it may not be measurable when running whole models. I think it's fine to merge as it is but good to keep in mind.